SistemaDiscusiones/requirements.txt
Django==1.6.2
South==0.8.4
Unipath==1.0
dj-database-url==0.3.0
dj-static==0.0.5
django-braces==1.4.0
django-redis-cache==0.11.1
django-toolbelt==0.0.1
djrill==0.9.0
gunicorn==18.0
oauthlib==0.6.1
psycopg2==2.5.2
pystache==0.5.3
python-openid==2.2.5
python-social-auth==0.1.22
redis==2.9.1
requests==2.2.1
requests-oauthlib==0.4.0
six==1.6.1
static==1.0.2
